Internals: Fix cppcheck warnings; no functional change intended

+++ src/V3List.h
@@ -60,7 +60,7 @@ private:
     T		m_nextp;	// Pointer to next element, NULL=end
     T		m_prevp;	// Pointer to previous element, NULL=beginning
     friend class V3List<T>;
-    static V3ListEnt* baseToListEnt(void* newbasep, uint32_t offset) {
+    static V3ListEnt* baseToListEnt(void* newbasep, size_t offset) {
 	// "this" must be a element inside of *basep
 	// Use that to determine a structure offset, then apply to the new base
 	// to get our new pointer information
@@ -80,7 +80,8 @@ public:
     // METHODS
     void pushBack (V3List<T>& listr, T newp) {
 	// "this" must be a element inside of *newp
-	uint32_t offset = (vluint8_t*)(this) - (vluint8_t*)(newp);
+	// cppcheck-suppress thisSubtraction
+	size_t offset = (vluint8_t*)(this) - (vluint8_t*)(newp);
 	m_nextp = NULL;
 	if (!listr.m_headp) listr.m_headp = newp;
 	m_prevp = listr.m_tailp;
@@ -89,7 +90,8 @@ public:
     }
     void pushFront (V3List<T>& listr, T newp) {
 	// "this" must be a element inside of *newp
-	uint32_t offset = (vluint8_t*)(this) - (vluint8_t*)(newp);
+	// cppcheck-suppress thisSubtraction
+	size_t offset = (vluint8_t*)(this) - (vluint8_t*)(newp);
 	m_nextp = listr.m_headp;
 	if (m_nextp) baseToListEnt(m_nextp,offset)->m_prevp = newp;
 	listr.m_headp = newp;
@@ -99,7 +101,8 @@ public:
     // Unlink from side
     void unlink (V3List<T>& listr, T oldp) {
 	// "this" must be a element inside of *oldp
-	uint32_t offset = (vluint8_t*)(this) - (vluint8_t*)(oldp);
+	// cppcheck-suppress thisSubtraction
+	size_t offset = (vluint8_t*)(this) - (vluint8_t*)(oldp);
 	if (m_nextp) baseToListEnt(m_nextp,offset)->m_prevp = m_prevp;
 	else listr.m_tailp = m_prevp;
 	if (m_prevp) baseToListEnt(m_prevp,offset)->m_nextp = m_nextp;
